requirements.txt
numpy==1.21.3
Pillow==8.3.2
torch==1.10.0
torchvision==0.11.1
